requirements.txt
scipy
soundfile
numpy
pyannote.audio
librosa